### 2023-10-27 #5991 @editorialbot
[PRE REVIEW]: SIRITVIS: Social Interaction Research Insights Topic Visualisation
**Submitting author:** @CodeEagle22 (<a href="http://orcid.org/0009-0004-9636-3611">Sagar Narwade</a>)
**Repository:** https://github.com/CodeEagle22/SIRITVIS/
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.3
