+++ README.md
@@ -14,8 +14,10 @@ One of the standout feature of SIRITVIS is its ability to map the spatial distri
 
 The innovative capabilities of this tool hold great potential in various domains, such as marketing, politics, and disaster management, empowering data-driven decision-making through spatial topic distribution insights. Organisations can leverage SIRITVIS to gain a deeper understanding of their customers and stakeholders, foster engagement, and facilitate informed decision-making processes based on comprehensive social media data analysis.
 
-## How to cite
-Narwade, S., Kant, G., Säfken, B., and Leiding, B. (2023), SIRITVIS: Social Interaction Research Insights Topic Visualisation.
+## 📝 How to cite
+
+Narwade, S., Kant, G., Säfken, B., and Leiding, B. (2023), SIRITVIS: Social Interaction Research Insights Topic Visualisation. Journal of Open Source Software, https://joss.theoj.org/papers/b51be70e9634e45d8035ee20b6147d76.
+
 
 
 ## Features
@@ -162,13 +164,13 @@ model_source = '../file/path/of/model.pkl'
 topic_mapper.TopicMapper(data_source, model_source)
 ```
 
-## Community guidelines
+## 📣 Community guidelines
 
 We encourage and welcome contributions to the SIRITVIS package. If you have any questions, want to report bugs, or have ideas for new features, please file an issue. 
 
 Additionally, we appreciate pull requests via GitHub. There are several areas where potential contributions can make a significant impact, such as enhancing the quality of topics in topic models when dealing with noisy data from Reddit, Instagram and Twitter or any external data sources, and improving the topic_mapper function to make it more interactive and independent from the notebook.
 
-## Authors
+## 📜 Authors
 
 - Sagar Narwade
 - Gillian Kant
