### 2023-10-30 #5991 @CodeEagle22
> Failed to discover a valid open source license

MIT License got uploaded on root directory
