### 2023-11-13 f390f6a
1.1.4 instastream.py assert
