+++ build/lib/SIRITVIS/topic_mapper.py
@@ -1,3 +1,6 @@
+# Copyright (c) [year] [your name]
+# This software is released under the MIT License.
+# https://opensource.org/licenses/MIT
 
 from ipywidgets import Button, Output
 import pandas as pd
