+++ build/lib/SIRITVIS/cleaner.py
@@ -1,3 +1,7 @@
+# Copyright (c) [year] [your name]
+# This software is released under the MIT License.
+# https://opensource.org/licenses/MIT
+
 import pickle
 import numpy as np
 import collections
