+++ SIRITVIS.egg-info/SOURCES.txt
@@ -1,3 +1,4 @@
+LICENSE
 README.md
 setup.py
 SIRITVIS/__init__.py
