+++ paper/paper.md
@@ -58,7 +58,7 @@ In recent years, the field of text data analysis from social media has witnessed
 
 Although alternatives such as TTLocVis [@Kant2020], TweetViz [@stojanovski2014] and Twitmo [@abuchmueller] have their merits, SIRITVIS sets itself apart by providing exceptional flexibility, usability, and comprehensiveness. Its extensive array of advanced topic models, alongside a distinctive capability for pinpointing geographical information and seamless integration with pyLDAvis [@sievert_shirley] for visualising outcomes, empowers researchers to extract profound insights from social media text data.
 
-What sets SIRITVIS apart is its comprehensive suite of evaluation metrics, facilitated by the octis tool [@terragni2020octis]. These metrics cover important aspects such as topic diversity, accuracy, inverted RBO, coherence, and Jaccard similarity, ensuring that the topic models generated by SIRITVIS are not only reliable, but also imbued with substantive meaning. This robust evaluation framework is a hallmark of the research, emphasizing the toolkit's dedication to producing meaningful results.
+What sets SIRITVIS apart is its comprehensive suite of evaluation metrics, facilitated by the octis tool [@terragni-etal-2021-octis]. These metrics cover important aspects such as topic diversity, accuracy, inverted RBO, coherence, and Jaccard similarity, ensuring that the topic models generated by SIRITVIS are not only reliable, but also imbued with substantive meaning. This robust evaluation framework is a hallmark of the research, emphasizing the toolkit's dedication to producing meaningful results.
 
 Furthermore, SIRITVIS places a strong emphasis on user-friendliness and accessibility. Its intuitive interface and detailed documentation cater to both novice and experienced users, making the toolkit approachable and easy to navigate. Additionally, SIRITVIS is designed to accommodate various data sources and formats, ensuring that researchers can adapt it to their specific research needs effortlessly. This commitment to accessibility broadens the toolkit's reach, making advanced text data analysis more accessible to a wider community of researchers and practitioners.
 
