+++ paper/paper.bib
@@ -7,39 +7,45 @@
 
 %% Saved with string encoding Unicode (UTF-8) 
 
-@inproceedings{terragni2020octis,
-    title={{OCTIS}: Comparing and Optimizing Topic Models is Simple!},
-    author={Terragni, Silvia and Fersini, Elisabetta and Galuzzi, Bruno Giovanni and Tropeano, Pietro and Candelieri, Antonio},
-    year={2021},
-    booktitle={Proceedings of the 16th Conference of the European Chapter of the Association for Computational Linguistics: System Demonstrations},
+@inproceedings{terragni-etal-2021-octis,
+    title = "{OCTIS}: Comparing and Optimizing Topic models is Simple!",
+    author = "Terragni, Silvia  and
+      Fersini, Elisabetta  and
+      Galuzzi, Bruno Giovanni  and
+      Tropeano, Pietro  and
+      Candelieri, Antonio",
+    editor = "Gkatzia, Dimitra  and
+      Seddah, Djam{\'e}",
+    booktitle = "Proceedings of the 16th Conference of the European Chapter of the Association for Computational Linguistics: System Demonstrations",
     month = apr,
     year = "2021",
+    address = "Online",
     publisher = "Association for Computational Linguistics",
-    url = "https://www.aclweb.org/anthology/2021.eacl-demos.31",
+    url = "https://aclanthology.org/2021.eacl-demos.31",
+    doi = "10.18653/v1/2021.eacl-demos.31",
     pages = "263--270",
+    abstract = "In this paper, we present OCTIS, a framework for training, analyzing, and comparing Topic Models, whose optimal hyper-parameters are estimated using a Bayesian Optimization approach. The proposed solution integrates several state-of-the-art topic models and evaluation metrics. These metrics can be targeted as objective by the underlying optimization procedure to determine the best hyper-parameter configuration. OCTIS allows researchers and practitioners to have a fair comparison between topic models of interest, using several benchmark datasets and well-known evaluation metrics, to integrate novel algorithms, and to have an interactive visualization of the results for understanding the behavior of each model. The code is available at the following link: \url{https://github.com/MIND-Lab/OCTIS}.",
 }
 
-@inproceedings{mehrotra2013improving,
-  title={Improving lda topic models for microblogs via tweet pooling and automatic labeling},
-  author={Mehrotra, Rishabh and Sanner, Scott and Buntine, Wray and Xie, Lexing},
-  booktitle={Proceedings of the 36th international ACM SIGIR conference on Research and development in information retrieval},
-  pages={889--892},
-  year={2013}
+@inproceedings{10.1145/2484028.2484166,
+author = {Mehrotra, Rishabh and Sanner, Scott and Buntine, Wray and Xie, Lexing},
+title = {Improving LDA Topic Models for Microblogs via Tweet Pooling and Automatic Labeling},
+year = {2013},
+isbn = {9781450320344},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/2484028.2484166},
+doi = {10.1145/2484028.2484166},
[…]
