+++ paper/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 # Summary
 
-SIRITVIS is a powerful text analysis tool that has been carefully designed to analyse data from Reddit, X (Twitter), Instagram, and any external text data sources. It utilises the latest tools to automatically extract information from these sources, clean the data, and optimise topic modeling more efficiently. The Package also contains data visualisation tools to improve the investigation and understanding of textual data. One feature of SIRITVIS is its ability to map social media posts on a global scale, connecting geographical locations with trending topics, thus providing valuable insights into the worldwide trends and conversations shaping our digital landscape. Additionally, it offers sentiment analysis on mapped data using the well-regarded VADER tool [@hutto2014vader]. This software is a valuable resource for the scientific community, offering deep insights into public discussions on various social platforms. It can be used for a wide range of research purposes, including the analysis of recent discussions on global geopolitical issues. Installing SIRITVIS is straightforward, as it can be accomplished using the pip package manager. Comprehensive installation instructions can typically be found in the dedicated repository of the package.^[https://github.com/CodeEagle22/SIRITVIS/]
+SIRITVIS is a powerful text analysis tool that has been carefully designed to analyse data from Reddit, X (Twitter), Instagram, and any external text data sources. It utilises the latest tools to automatically extract information from these sources, clean the data, and optimise topic modeling more efficiently. The Package also contains data visualisation tools to improve the investigation and understanding of textual data. One feature of SIRITVIS is its ability to map social media posts on a global scale, connecting geographical locations with trending topics, thus providing valuable insights into the worldwide trends and conversations shaping our digital landscape. Additionally, it offers sentiment analysis on mapped data using the well-regarded VADER tool [@Hutto_Gilbert_2014]. This software is a valuable resource for the scientific community, offering deep insights into public discussions on various social platforms. It can be used for a wide range of research purposes, including the analysis of recent discussions on global geopolitical issues. Installing SIRITVIS is straightforward, as it can be accomplished using the pip package manager. Comprehensive installation instructions can typically be found in the dedicated repository of the package.^[https://github.com/CodeEagle22/SIRITVIS/]
 
 
 ## Statement of Need
@@ -48,7 +48,7 @@ SIRITVIS encompasses a comprehensive suite of functionalities designed for data 
 
 In addition, the package includes an evaluation module, which allows for the assessment of trained models using a variety of metrics tailored to the user's specific requirements. Furthermore, the software offers the capability to analyse and retrieve the most proficiently trained models, further enhancing its utility and practicality for scientific research and analysis.
 
-To enrich the analysis and comprehension of textual data, SIRITVIS includes two powerful data visualisation tools: PyLDAvis [@sievert_shirley] and Word Cloud [@Mueller_Wordcloud_2023]. The useful data visualisation tool PyLDAvis, created by Sievert and Shirley, improves the interpretation of topic models from textual data. In order to give users a deeper and more intuitive understanding of the latent themes contained in the text corpus, it creates dynamic and interactive visualisations that assist users in exploring the connections between subjects and their associated keywords (see figure 1). On the other hand, the Word Cloud tool offers an engaging and straightforward means of visually representing the most frequently occurring words in the dataset, simplifying the identification of crucial keywords and patterns at a glance (see figure 2). Together, these tools provide users with a comprehensive and user-friendly approach to uncover and communicate the hidden patterns and insights within their textual data.
[…]
