+++ paper/paper.bib
@@ -143,7 +143,10 @@ year = {2023}
 	url = {https://github.com/abuchmueller/Twitmo},
 }
 
-@article{Hutto_Gilbert_2014, title={VADER: A Parsimonious Rule-Based Model for Sentiment Analysis of Social Media Text}, volume={8}, url={https://ojs.aaai.org/index.php/ICWSM/article/view/14550}, DOI={10.1609/icwsm.v8i1.14550}, abstractNote={ &lt;p&gt; The inherent nature of social media content poses serious challenges to practical applications of sentiment analysis. We present VADER, a simple rule-based model for general sentiment analysis, and compare its effectiveness to eleven typical state-of-practice benchmarks including LIWC, ANEW, the General Inquirer, SentiWordNet, and machine learning oriented techniques relying on Naive Bayes, Maximum Entropy, and Support Vector Machine (SVM) algorithms. Using a combination of qualitative and quantitative methods, we first construct and empirically validate a gold-standard list of lexical features (along with their associated sentiment intensity measures) which are specifically attuned to sentiment in microblog-like contexts. We then combine these lexical features with consideration for five general rules that embody grammatical and syntactical conventions for expressing and emphasizing sentiment intensity. Interestingly, using our parsimonious rule-based model to assess the sentiment of tweets, we find that VADER outperforms individual human raters (F1 Classification Accuracy = 0.96 and 0.84, respectively), and generalizes more favorably across contexts than any of our benchmarks. &lt;/p&gt; }, number={1}, journal={Proceedings of the International AAAI Conference on Web and Social Media}, author={Hutto, C. and Gilbert, Eric}, year={2014}, month={May}, pages={216-225} }
+
+@article{Hutto_Gilbert_2014, 
+title={VADER: A Parsimonious Rule-Based Model for Sentiment Analysis of Social Media Text}, volume={8}, url={https://ojs.aaai.org/index.php/ICWSM/article/view/14550}, DOI={10.1609/icwsm.v8i1.14550}, abstractNote={ &lt;p&gt; The inherent nature of social media content poses serious challenges to practical applications of sentiment analysis. We present VADER, a simple rule-based model for general sentiment analysis, and compare its effectiveness to eleven typical state-of-practice benchmarks including LIWC, ANEW, the General Inquirer, SentiWordNet, and machine learning oriented techniques relying on Naive Bayes, Maximum Entropy, and Support Vector Machine (SVM) algorithms. Using a combination of qualitative and quantitative methods, we first construct and empirically validate a gold-standard list of lexical features (along with their associated sentiment intensity measures) which are specifically attuned to sentiment in microblog-like contexts. We then combine these lexical features with consideration for five general rules that embody grammatical and syntactical conventions for expressing and emphasizing sentiment intensity. Interestingly, using our parsimonious rule-based model to assess the sentiment of tweets, we find that VADER outperforms individual human raters (F1 Classification Accuracy = 0.96 and 0.84, respectively), and generalizes more favorably across contexts than any of our benchmarks. &lt;/p&gt; }, number={1}, journal={Proceedings of the International AAAI Conference on Web and Social Media}, author={Hutto, C. and Gilbert, Eric}, year={2014}, month={May}, pages={216-225}
+}
 
 
 
