+++ paper/paper.md
@@ -56,7 +56,7 @@ SIRITVIS is renowned for its remarkable ability to map the spatial distribution 
 
 In recent years, the field of text data analysis from social media has witnessed remarkable advancements, offering researchers and practitioners an array of toolkits and approaches to delve into the wealth of online content. Within this dynamic landscape, it becomes imperative to discern the distinctive features of our research, encapsulated in the SIRITVIS framework, as it stands apart from existing related work.
 
-Although alternatives such as TTLocVis [@Kant2020], TweetViz [@stojanovski2014] and Twitmo [@abuchmueller] have their merits, SIRITVIS sets itself apart by providing exceptional flexibility, usability, and comprehensiveness. Its extensive array of advanced topic models, alongside a distinctive capability for pinpointing geographical information and seamless integration with pyLDAvis [@sievert_shirley] for visualising outcomes, empowers researchers to extract profound insights from social media text data.
+Although alternatives such as TTLocVis [@Kant2020], TweetViz [@stojanovski2014] and Twitmo [@abuchmueller] have their merits, SIRITVIS sets itself apart by providing exceptional flexibility, usability, and comprehensiveness. Its extensive array of advanced topic models, alongside a distinctive capability for pinpointing geographical information and seamless integration with pyLDAvis [@sievert-shirley-2014-ldavis] for visualising outcomes, empowers researchers to extract profound insights from social media text data.
 
 What sets SIRITVIS apart is its comprehensive suite of evaluation metrics, facilitated by the octis tool [@terragni-etal-2021-octis]. These metrics cover important aspects such as topic diversity, accuracy, inverted RBO, coherence, and Jaccard similarity, ensuring that the topic models generated by SIRITVIS are not only reliable, but also imbued with substantive meaning. This robust evaluation framework is a hallmark of the research, emphasizing the toolkit's dedication to producing meaningful results.
 
