### 2024-01-18 #5991 @cjbarrie
Hi All. Sorry for slowness. I'm happy to do this. One Q before I begin: not sure how I would go about testing the Twitter  connection now the the API is paid?
