### 2024-01-19 #5991 @julianmcauley
unsubscribe

On Fri, 19 Jan 2024 at 08:38, The Open Journals editorial robot <
***@***.***> wrote:

> Closed #5991 <https://github.com/openjournals/joss-reviews/issues/5991>
> as completed.
>
