### 2024-01-22 #5991 @CodeEagle22
> Hi All. Sorry for slowness. I'm happy to do this. One Q before I begin: not sure how I would go about testing the Twitter connection now the the API is paid?

Hello @cjbarrie, Thank you for bringing this to my attention. In my previous assessment, the Twitter streaming feature's API was free and operational. However, the current cost of approximately [$5000 per month](https://developer.twitter.com/en/docs/twitter-api/getting-started/about-twitter-api) as you mentioned for utilizing the API with multiple filters which unaffordable for me. I am unsure about your testing approach, but unless an alternative solution is identified, I may have to consider removing the reference to this feature from the paper.
