### 2024-02-01 #6243 @cjbarrie
Hi All,

I've been going through this and am bumping up against quite a few errors. I detail these in commented lines below:

```python
from SIRITVIS import twitter_streamer, insta_streamer, reddit_streamer, cleaner, topic_model, topic_visualise, topic_mapper
import ssl
import nltk
try:
    _create_unverified_https_context = ssl._create_unverified_context
except AttributeError:
    pass
else:
    ssl._create_default_https_context = _create_unverified_https_context
nltk.download('stopwords')

# Run the streaming process to retrieve raw data based on the specified keywords
client_id = "******************"
client_secret = "**********************"
user_agent = "**************"
keywords = ['and'] # just to get random data
save_path = 'data/'
raw_data = reddit_streamer.RedditStreamer(client_id,client_secret,user_agent,save_path,keywords).run()

cleaner_obj = cleaner.Cleaner(data_source='data/',data_source_type='reddit')
cleaned_file = cleaner_obj.saving('data_cleaned/',data_save_name='reddit_cleaned')

cleaned_file.to_csv('corpus.csv', index=False, header=True)

# gives error if you pass cleaned_file to dataset_source
model = topic_model.TopicModeling(num_topics=10, dataset_source='corpus.csv',
learning_rate=0.001, batch_size=32, activation='softplus', num_layers=3, num_neurons=100,
dropout=0.2, num_epochs=100, save_model=False, model_path=None, train_model='NeuralLDA',
evaluation=['topicdiversity','invertedrbo','jaccardsimilarity'])

saved_model = model.run()
# gives nltk install error unless you use SSL routine above

# gives following error: An error occurred: max_df corresponds to < documents than min_df
vis_model = topic_visualise.PyLDAvis(data_source='corpus.csv',num_topics=5,text_column='text')
vis_model.visualize()

vis_model = topic_visualise.Wordcloud(data_source='corpus.csv',text_column='text')
vis_model.visualize() #nothing happens when trying to visualize
```
