### 2024-02-05 #6243 @SagarNarwade
Hello @cjbarrie,

Thank you for providing your feedback. I'm pleased to inform you that the issues you mentioned have been addressed in the latest version (1.1.6). To ensure optimal plotting, I suggest utilizing an Integrated Development Environment (IDE) as some plots may encounter difficulties saving as image or HTML files when executing code via command line or terminal. Additionally, for optimal results or visualization appearance, it is recommended to maintain a minimum csv file size of 500KB to 1MB.
