### 2024-02-15 #6243 @cjbarrie
Hi @CodeEagle22, this version now allows the cleaned_file object to be passed to the model specification function. But every time I run this, I get the following error:

```python

model = topic_model.TopicModeling(num_topics=10, dataset_source=cleaned_file,
learning_rate=0.001, batch_size=32, activation='softplus', num_layers=3, num_neurons=100,
dropout=0.2, num_epochs=100, save_model=False, model_path=None, train_model='NeuralLDA',
evaluation=['topicdiversity','invertedrbo','jaccardsimilarity'])

```

```
ValueError: With n_samples=0, test_size=0.15 and train_size=None, the resulting train set will be empty. Adjust any of the aforementioned parameters.
```

I _think_ this is because the actual reddit data I've collected is not sufficiently large. But the raw data I collected is >2GB in size. When I generate the cleaned file with:

```python
cleaner_obj = cleaner.Cleaner(data_source='data/',data_source_type='reddit')
cleaned_file = cleaner_obj.saving('data_cleaned/',data_save_name='reddit_cleaned')
```

I get a .csv file with just 15 rows every time. Is there an argument I need to add to make sure that all of the data is combined with this function? It would be helpful to have more documentation so I could answer this myself.
