### 2024-02-15 #6243 @CodeEagle22
Hey @cjbarrie,

Regarding the issue you're facing with the cleaned_file object in the model specification function, it seems that the error is due to an empty training set. This could be because the cleaned file generated from the raw Reddit data only contains 15 rows, which might not be sufficient for training the model effectively.

Multiple filters have been applied during the cleaning process to remove spam posts. While this ensures data quality, it might result in a reduced number of rows in the cleaned file. If you use given [sample raw data](https://github.com/CodeEagle22/SIRITVIS/tree/main/sample_dataset/Reddit/raw_data) of reddit you will find it generates thousands of lines in clean file which means in your raw streaming data there might be very less ham posts (15 rows).

To address this, I suggest considering the following points:

1. Diversify Keywords: When collecting streaming data, employing a broader range of keywords can help diversify the dataset. For instance, incorporating terms like 'and', 'oil', and 'energy' could broaden the scope of the collected data, potentially yielding a larger pool of usable content for cleaning and analysis.

2. Ensure Adequate File Size: The cleaned file's size is notably small, which is problematic for effective training. It's advisable to aim for a more substantial file size, ideally exceeding 500 KB.

By addressing these points, you can mitigate the issue of an empty training set and ensure that the model specification function receives a sufficiently large and diverse dataset for effective training.
