### 2024-02-20 #6243 @cjbarrie
Thanks a lot for this, @CodeEagle22. This is helpful. I can use the sample data for tests... One thing: none of this seems to have been documented anywhere. It's hard for us to review when we don't know exactly what a function is doing without documentation (absent going into all of the source code etc.). And this would also make the library a lot more useful for applied users. I strongly recommend that you consider providing documentation for the functions, detailing the different variables and parameters. My apologies if this exists and I just haven't found it
