### 2024-02-21 #6243 @cjbarrie
Hi @CodeEagle22, I've returned to this I'm now using the sample_dataset rather than collecting my own data. When I do this (in VS Code). I get the following error:

```python

# gives following error: The visualisation is based on Latent Dirichlet Allocation (LDA) model.
# An error occurred: 'NoneType' object has no attribute 'display_formatter'
vis_model = topic_visualise.PyLDAvis(data_source=cleaned_file,num_topics=5,text_column='text')
vis_model.visualize()

# gives following error: The visualisation is based on Latent Dirichlet Allocation (LDA) model.
# An error occurred: 'NoneType' object has no attribute 'display_formatter'
vis_model = topic_visualise.Wordcloud(data_source=cleaned_file,text_column='text')
vis_model.visualize()
```

I therefore decided to try out a different IDE and used a Google Colab notebook. When I did this, I got an import error related to pyLDAvis:

```python
from SIRITVIS import twitter_streamer, insta_streamer, reddit_streamer, cleaner, topic_model, topic_visualise, topic_mapper
import nltk
nltk.download('stopwords')
```

```
/usr/local/lib/python3.10/dist-packages/ipykernel/ipkernel.py:283: DeprecationWarning: `should_run_async` will not call `transform_cell` automatically in the future. Please pass the result to `transformed_cell` argument and any exception that happen during thetransform in `preprocessing_exc_tuple` in IPython 7.17 and above.
  and should_run_async(code)
ImportError                               Traceback (most recent call last)
[<ipython-input-5-4aab91ff331f>](https://localhost:8080/#) in <cell line: 1>()
----> 1 from SIRITVIS import twitter_streamer, insta_streamer, reddit_streamer, cleaner, topic_model, topic_visualise, topic_mapper
      2 import nltk
      3 nltk.download('stopwords')

1 frames
[/usr/local/lib/python3.10/dist-packages/SIRITVIS/topic_visualise.py](https://localhost:8080/#) in <module>
     15 from IPython.core.display import display, HTML
     16 import pyLDAvis
---> 17 from pyLDAvis.lda_model import LDA
     18 import warnings
     19 from wordcloud import WordCloud

ImportError: cannot import name 'LDA' from 'pyLDAvis.lda_model' (/usr/local/lib/python3.10/dist-packages/pyLDAvis/lda_model.py)

```

So I decided to try jupyter lab. When I did this, all of the code ran but there was again no visualization output.

I think at this stage what I need is a notebook that I can open and run. All of these errors make me a bit nervous and will also mean it is unlikely that your end user is going to have much luck either.
