+++ README.md
@@ -22,11 +22,11 @@ The innovative capabilities of this tool hold great potential in various domains
 
 Narwade, S., Kant, G., Säfken, B., and Leiding, B. (2023), SIRITVIS: Social Interaction Research Insights Topic Visualisation. Journal of Open Source Software, https://joss.theoj.org/papers/b51be70e9634e45d8035ee20b6147d76.
 
-## Requirements
+## Advisory
 
-- Python 3.10.*
-- An Integrated Development Environment (IDE) like Visual Studio Code or Jupyter Notebook for better plot visualization.
-
+- Ensure Python version 3.10 or higher.
+- Utilize IDEs like Visual Studio or platforms like Google Colab for enhanced plot visualization.
+- Refer to the provided sample dataset for better comprehension.
 
 ## 💡 Features
 
@@ -55,6 +55,7 @@ from SIRITVIS import twitter_streamer, insta_streamer, reddit_streamer, cleaner,
 ```
 
 ### Streaming Reddit Data
+- For authentication with the Reddit Streaming API, follow the steps outlined in this [tutorial](https://towardsdatascience.com/how-to-use-the-reddit-api-in-python-5e05ddfd1e5c).
 
 ```python
 # Run the streaming process to retrieve raw data based on the specified keywords
@@ -62,12 +63,13 @@ from SIRITVIS import twitter_streamer, insta_streamer, reddit_streamer, cleaner,
 client_id = "XXXXXXXXXX"
 client_secret = "XXXXXXXXX"
 user_agent = "XXXXXXXXXX"
-keywords = ['Specific','Keywords'] # default is None
+keywords = ['Specific','Keywords'] # default is None # Use multiple keywords for a more varied dataset during streaming data collection.
 save_path = '../folder/path/to/store/the/data/'
 raw_data = reddit_streamer.RedditStreamer(client_id,client_secret,user_agent,save_path,keywords).run()
 ```
 
 ### Streaming Twitter Data
+- For authentication with the Twitter Streaming API, follow the page [Twitter Developer Account](https://developer.twitter.com/en/docs/twitter-api/getting-started/about-twitter-api)
 
 ```python
 # Run the streaming process to retrieve raw data based on the specified keywords and for specific location
@@ -83,7 +85,7 @@ locations = [51.416016,5.528511,90.966797,34.669359] # box coordinates.
 languages = ['en', 'es'] 
 
 # Keywords to track
-keywords = ['Specific','Keywords'] # default is None
+keywords = ['Specific','Keywords'] # default is None # Use multiple keywords for a more varied dataset during streaming data collection.
 
 # Save path for collected data
 save_path = '../folder/path/to/store/the/data/'
@@ -102,6 +104,7 @@ raw_data = twitter_streamer.TwitterStreamer(
 ```
 
 ### Streaming Instagram Data
+- For authentication with the Instagram Streaming API, sign up the page [apify](https://apify.com/apify/instagram-hashtag-scraper)
 
 ```python
 # Run the streaming process to retrieve raw data based on the specified keywords
@@ -111,7 +114,7 @@ api_token = 'apify_api_XXXXXXXXX'
 save_path = '../folder/path/to/store/the/data/'
 instagram_username = 'XXXXXXXXX'
 instagram_password = 'XXXXXXXXX'
-hashtags = ['Specific','Keywords'] # default is ['instagram']
+hashtags = ['Specific','Keywords'] # default is ['instagram'] # Use multiple keywords for a more varied dataset during streaming data collection.
 limit =  20 # number of post captions to extract. default is 100
[…]
