### 2024-02-25 #6243 @CodeEagle22
Hey @cjbarrie,

Sorry for inconvenience. I've updated the issue and provided the latest documentation in pip installing SIRITVIS==1.1.7.1. This should address all the issues and make it more robust for end users.
