### 2024-03-18 #6243 @cjbarrie
Hiya. I just went back to test with the new version of the library that @CodeEagle22 mentioned. I'm still bumping into a lot of errors, unfortunately. As I said above, it would really help if the authors could provide a Colab runthrough of the working library. As written, I just don't have the time to keep testing code that is not functioning as described. This is the fourth or fifth time I've been back to it.

I appreciate a lot of work has gone into this software. But realistically, the take-up by any end user will be minimal without further guidance on functionality and without resolving the errors I keep encountering.

Here is my code, with annotations of the errors this routine throws:

```python
from SIRITVIS import twitter_streamer, insta_streamer, reddit_streamer, cleaner, topic_model, topic_visualise, topic_mapper

import ssl
try:
    _create_unverified_https_context = ssl._create_unverified_context
except AttributeError:
    pass
else:
    ssl._create_default_https_context = _create_unverified_https_context

import nltk
nltk.download('stopwords')

# using the Github sample_dataset data
cleaner_obj = cleaner.Cleaner(data_source='data/',data_source_type='reddit')
cleaned_file = cleaner_obj.saving('data_cleaned/',data_save_name='reddit_cleaned')

cleaned_file.to_csv('corpus.csv', index=False, header=True)

# This no longer takes "cleaned_file" object as parameter
model = topic_model.TopicModeling(num_topics=10, dataset_source='data_cleaned/reddit_cleaned_part_1.csv',
learning_rate=0.001, batch_size=32, activation='softplus', num_layers=3, num_neurons=100,
dropout=0.2, num_epochs=100, save_model=False, model_path=None, train_model='NeuralLDA',
evaluation=['topicdiversity','invertedrbo','jaccardsimilarity'])

saved_model = model.run()

# gives following error: The visualisation is based on Latent Dirichlet Allocation (LDA) model.
# An error occurred: 'NoneType' object has no attribute 'display_formatter'
vis_model = topic_visualise.PyLDAvis(data_source=cleaned_file,num_topics=5,text_column='text')
vis_model.visualize()

# gives following error: An error occurred: 'ImageDraw' object has no attribute 'textsize'
vis_model = topic_visualise.Wordcloud(data_source=cleaned_file,text_column='text')
vis_model.visualize()
```
