### 2024-03-19 #6243 @n3mo
I'm running into similar errors as @cjbarrie. The documentation suggests that the package is optimized for python notebooks, but it seems more accurate to say that it currently depends on it. I encountered the same two errors shared by @cjbarrie. The 'display_formatter' error associated with `topic_visualize.PyLDAvis()` does go away when model results are visualized in a Jupyter notebook. I am able to successfully create the intertopic distance map visualization.

It seems that the PyLDAvis package that does the heavy lifting here offers a `pyLDAvis.save_html()` method, which I believe allows visualization support outside of notebook contexts (by saving directly to file). However, SIRITVIS either doesn't make this available to it's visualizer, or I cannot figure out how to find it. Either way, @cjbarrie's observation is accurate: this is a significant barrier to new user take-up.

Aside from the above, the 'textsize' error associated with `topic_visualise.Wordcloud()` persists in both notebooks and interactive python shells, so I have been unable to generate this figure.
