### 2024-03-19 #6243 @n3mo
I can confirm that the TopicMapper visualization also works for me, but also only in a notebook context. There is no error in an interactive python shell context--instead, an IPython.core.display.HTML object is returned. I believe this can be saved to file for display in a browser. But clearly, a notebook context is preferred here as well.
