### 2024-05-05 #6243 @CodeEagle22
Hello @cjbarrie and @n3mo,

You'll discover the latest version 1.1.8 along with a [notebook example](https://github.com/CodeEagle22/SIRITVIS/blob/main/test_notebook/Demo_Jupyter_Notebook.ipynb). I've included code to remove SSL, enabling access to URLs starting with "http://". Additionally, the cleaned_file parameter is now accepted by topic_model.TopicModeling(). Both topic_visualise() functionalities work seamlessly, as demonstrated in the [sample notebook.](https://github.com/CodeEagle22/SIRITVIS/blob/main/test_notebook/Demo_Jupyter_Notebook.ipynb).
The pyLDAvis.save_html() function isn't accessible in this version as it creates dependencies on other libraries.

[Colab-Notebook](https://colab.research.google.com/drive/1JFMD00m1-ofBG6Z2fMJJXU41zEKJ5H5j?usp=sharing)
