### 2024-05-19 #6243 @cjbarrie
Hi All,

Thanks for all the improvements you've made to this, @CodeEagle22. That makes use a lot simpler and will likely result in more people using the software.

I'm nearly ready to sign off but two outstanding things remain:

1. There is no explanation given of how this software will work with Twitter now that the API is no longer available to users (unless paying huge sums for it). I am confused why the authors continue to sell this as a tool compatible with Twitter.
2. The language in the article is a little overloaded at point e.g., "SIRITVIS is renowned for its remarkable ability..." At various points in the text the authors are overdoing it with the language. I would ask that they consider toning down the language in the article. This is supposed to explain the core functionality of the software not sell it to potential investors.
