+++ paper_old/paper.bib
@@ -1,0 +1,154 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% http://bibdesk.sourceforge.net/
+
+
+%% Created for Christoph at 2020-08-12 17:11:07 +0200 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+@inproceedings{terragni-etal-2021-octis,
+    title = "{OCTIS}: Comparing and Optimizing Topic models is Simple!",
+    author = "Terragni, Silvia  and
+      Fersini, Elisabetta  and
+      Galuzzi, Bruno Giovanni  and
+      Tropeano, Pietro  and
+      Candelieri, Antonio",
+    editor = "Gkatzia, Dimitra  and
+      Seddah, Djam{\'e}",
+    booktitle = "Proceedings of the 16th Conference of the European Chapter of the Association for Computational Linguistics: System Demonstrations",
+    month = apr,
+    year = "2021",
+    address = "Online",
+    publisher = "Association for Computational Linguistics",
+    url = "https://aclanthology.org/2021.eacl-demos.31",
+    doi = "10.18653/v1/2021.eacl-demos.31",
+    pages = "263--270",
+    abstract = "In this paper, we present OCTIS, a framework for training, analyzing, and comparing Topic Models, whose optimal hyper-parameters are estimated using a Bayesian Optimization approach. The proposed solution integrates several state-of-the-art topic models and evaluation metrics. These metrics can be targeted as objective by the underlying optimization procedure to determine the best hyper-parameter configuration. OCTIS allows researchers and practitioners to have a fair comparison between topic models of interest, using several benchmark datasets and well-known evaluation metrics, to integrate novel algorithms, and to have an interactive visualization of the results for understanding the behavior of each model. The code is available at the following link: \url{https://github.com/MIND-Lab/OCTIS}.",
+}
+
+@inproceedings{10.1145/2484028.2484166,
+author = {Mehrotra, Rishabh and Sanner, Scott and Buntine, Wray and Xie, Lexing},
+title = {Improving LDA Topic Models for Microblogs via Tweet Pooling and Automatic Labeling},
+year = {2013},
+isbn = {9781450320344},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/2484028.2484166},
+doi = {10.1145/2484028.2484166},
+abstract = {Twitter, or the world of 140 characters poses serious challenges to the efficacy of topic models on short, messy text. While topic models such as Latent Dirichlet Allocation (LDA) have a long history of successful application to news articles and academic abstracts, they are often less coherent when applied to microblog content like Twitter. In this paper, we investigate methods to improve topics learned from Twitter content without modifying the basic machinery of LDA; we achieve this through various pooling schemes that aggregate tweets in a data preprocessing step for LDA. We empirically establish that a novel method of tweet pooling by hashtags leads to a vast improvement in a variety of measures for topic coherence across three diverse Twitter datasets in comparison to an unmodified LDA baseline and a variety of pooling schemes. An additional contribution of automatic hashtag labeling further improves on the hashtag pooling results for a subset of metrics. Overall, these two novel schemes lead to significantly improved LDA topic models on Twitter content.},
+booktitle = {Proceedings of the 36th International ACM SIGIR Conference on Research and Development in Information Retrieval},
+pages = {889–892},
+numpages = {4},
+keywords = {topic modeling, microblogs, lda},
[…]
