+++ paper_old/paper.md
@@ -35,14 +35,14 @@ bibliography: paper.bib
 
 # Summary
 
-SIRITVIS is a comprehensive text analysis tool developed to analyze data from Reddit, Instagram, and various external text data sources. It utilizes advanced methodologies to extract information, clean data, and optimize topic modeling efficiently. The package includes data visualization tools that assist in exploring and understanding textual data. A notable feature of SIRITVIS is its ability to map social media posts globally, correlating geographical locations with trending topics, thus providing insights into global trends and digital conversations. It also performs sentiment analysis on mapped data using the VADER tool [@Hutto_Gilbert_2014]. This software is a valuable resource for the scientific community, enabling the exploration of public discussions across multiple social platforms. It supports a range of research purposes, including the analysis of current discussions on global geopolitical issues. The installation of SIRITVIS is straightforward and facilitated by the pip package manager. Detailed installation instructions are available in the package's repository^[https://github.com/CodeEagle22/SIRITVIS/].
+**SIRITVIS** is a comprehensive text analysis tool developed to analyze data from Reddit, Instagram, and various external text data sources. It utilizes advanced methodologies to extract information, clean data, and optimize topic modeling efficiently. The package includes data visualization tools that assist in exploring and understanding textual data. A notable feature of SIRITVIS is its ability to map social media posts globally, correlating geographical locations with trending topics, thus providing insights into global trends and digital conversations. It also performs sentiment analysis on mapped data using the VADER tool [@Hutto_Gilbert_2014]. This software is a valuable resource for the scientific community, enabling the exploration of public discussions across multiple social platforms. It supports a range of research purposes, including the analysis of current discussions on global geopolitical issues. The installation of SIRITVIS is straightforward and facilitated by the pip package manager. Detailed installation instructions are available in the package's repository^[https://github.com/CodeEagle22/SIRITVIS/].
 
 ## Statement of Need
 
 
 The proliferation of social media platforms has significantly changed how we communicate, share information, and express opinions on a variety of topics. Platforms like Reddit and Instagram serve as major hubs for public discussion. Analyzing text data from these platforms can provide insights into public sentiments, preferences, and trending discussions, which are valuable for areas such as marketing, politics, and disaster management.
 
-Handling the large volume of unstructured text data from social media can be challenging due to its dynamic and expansive nature. To help with this, we present SIRITVIS, an open-source text analysis package designed to facilitate the exploration and analysis of social media data. SIRITVIS incorporates several advanced neural topic models, including AVITM [@srivastava2017autoencoding], as well as other widely used models like Latent Dirichlet Allocation (LDA) [@blei2003latent], Neural Latent Dirichlet Allocation (NeuralLDA) [@srivastava2017autoencoding], Product Latent Dirichlet Allocation (ProdLDA) [@srivastava2017autoencoding], and Contextualized Topic Models (CTM) [@bianchi-etal-2021-cross]. These models help in automatically identifying and extracting key topics in an unsupervised manner, allowing users to efficiently explore large text datasets and uncover meaningful patterns.
[…]
