+++ SIRITVIS.egg-info/SOURCES.txt
@@ -8,7 +8,6 @@ SIRITVIS/reddit_streamer.py
 SIRITVIS/topic_mapper.py
 SIRITVIS/topic_model.py
 SIRITVIS/topic_visualise.py
-SIRITVIS/twitter_streamer.py
 SIRITVIS.egg-info/PKG-INFO
 SIRITVIS.egg-info/SOURCES.txt
 SIRITVIS.egg-info/dependency_links.txt
