### 2024-06-10 #6243 @CodeEagle22
Hi @cjbarrie ,

It seems that Twitter has blocked all free streaming services, leaving us with no other choice but to rely on their costly API for data scraping. Consequently, I have decided to remove the Twitter streaming feature from the latest version of the package (2.0.0). Additionally, I have updated the text of the paper to make it more neutral.

Thank you!
