### 2024-06-18 #6243 @CodeEagle22
Hello @n3mo,

I’ve updated to version 2.0.0. Are there any improvements you'd like me to make?
