### 2024-06-29 #6243 @arfon
> It seems the review checklist is completed. Could you please inform me about the next steps for finalising the paper publication?

:wave: folks. JOSS EiC here. I'm going to help Olivia out here with this paper.

/ cc @samhforbes
