### 2024-07-13 #6243 @CodeEagle22
Hello @arfon,

I have reviewed everything once more. The software version 2.0.0 is functioning correctly and includes the changes resulting from this review. The paper is thorough and verified. I have uploaded the archived version on Figshare, and here is the DOI: [10.6084/m9.figshare.26298487](https://figshare.com/s/25fdf67037cc1ae08553).

Please let me know if anything else is required or needs to be updated.

Thank you
