### 2024-07-18 #6243 @xillig
Dear @arfon and @samhforbes,

I am Gillian Kant, one of the co-authors of the paper. We deeply appreciate the effort you have already invested in advancing this project towards publication. Please let us know if there is any way we can assist you further to expedite the process.

Thank you!
