### 2024-07-19 #6243 @samhforbes
Hi @CodeEagle22 this looks good so far. The only thing I'd ask is to clean up consistency in the capitalisation of titles in the references, otherwise the paper reads clearly. Can you let me know when this is fixed?
