+++ paper/paper.bib
@@ -1,14 +1,5 @@
-%% This BibTeX bibliography file was created using BibDesk.
-%% http://bibdesk.sourceforge.net/
-
-
-%% Created for Christoph at 2020-08-12 17:11:07 +0200 
-
-
-%% Saved with string encoding Unicode (UTF-8) 
-
 @inproceedings{terragni-etal-2021-octis,
-    title = "{OCTIS}: Comparing and Optimizing Topic models is Simple!",
+    title = "{OCTIS}: Comparing and Optimizing Topic Models Is Simple!",
     author = "Terragni, Silvia  and
       Fersini, Elisabetta  and
       Galuzzi, Bruno Giovanni  and
@@ -45,10 +36,8 @@ location = {Dublin, Ireland},
 series = {SIGIR '13}
 }
 
-
-
 @misc{srivastava2017autoencoding,
-      title={Autoencoding Variational Inference For Topic Models}, 
+      title={Autoencoding Variational Inference for Topic Models}, 
       author={Akash Srivastava and Charles Sutton},
       year={2017},
       eprint={1703.01488},
@@ -61,12 +50,10 @@ series = {SIGIR '13}
 	Booktitle = {Proceedings of the Data Mining and Data Warehouses},
 	Date-Modified = {2020-08-12 17:10:23 +0200},
 	Pages = {1--4},
-	Title = {Tweetviz: Twitter data visualization},
+	Title = {Tweetviz: Twitter Data Visualization},
 	Year = {2014}}
 
 @article{Kant2020, doi = {10.21105/joss.02507}, url = {https://doi.org/10.21105/joss.02507}, year = {2020}, publisher = {The Open Journal}, volume = {5}, number = {54}, pages = {2507}, author = {Gillian Kant and Christoph Weisser and Benjamin Säfken}, title = {TTLocVis: A Twitter Topic Location Visualization Package}, journal = {Journal of Open Source Software} }
-
-
 
 @article{blei2003latent,
   title={Latent Dirichlet Allocation},
@@ -87,7 +74,7 @@ year = {2023}
 }
 
 @inproceedings{sievert-shirley-2014-ldavis,
-    title = "{LDA}vis: A method for visualizing and interpreting topics",
+    title = "{LDA}vis: A Method for Visualizing and Interpreting Topics",
     author = "Sievert, Carson  and
       Shirley, Kenneth",
     editor = "Chuang, Jason  and
@@ -106,7 +93,7 @@ year = {2023}
 }
 
 @inproceedings{bianchi-etal-2021-cross,
-    title = "Cross-lingual Contextualized Topic Models with Zero-shot Learning",
+    title = "Cross-Lingual Contextualized Topic Models with Zero-Shot Learning",
     author = "Bianchi, Federico  and
       Terragni, Silvia  and
       Hovy, Dirk  and
@@ -123,32 +110,24 @@ year = {2023}
     abstract = "Many data sets (e.g., reviews, forums, news, etc.) exist parallelly in multiple languages. They all cover the same content, but the linguistic differences make it impossible to use traditional, bag-of-word-based topic models. Models have to be either single-language or suffer from a huge, but extremely sparse vocabulary. Both issues can be addressed by transfer learning. In this paper, we introduce a zero-shot cross-lingual topic model. Our model learns topics on one language (here, English), and predicts them for unseen documents in different languages (here, Italian, French, German, and Portuguese). We evaluate the quality of the topic predictions for the same document in different languages. Our results show that the transferred topics are coherent and stable across languages, which suggests exciting future research directions.",
 }
 
-
 @article{albalawi-2020,
 	author = {Albalawi, Rania and Yeap, Tet Hin and Benyoucef, Morad},
[…]
