+++ paper/paper.md
@@ -35,34 +35,39 @@ bibliography: paper.bib
 
 # Summary
 
-SIRITVIS is a comprehensive text analysis tool developed to analyze data from Reddit, Instagram, and various external text data sources. It utilizes advanced methodologies to extract information, clean data, and optimize topic modeling efficiently. The package includes data visualization tools that assist in exploring and understanding textual data. A notable feature of SIRITVIS is its ability to map social media posts globally, correlating geographical locations with trending topics, thus providing insights into global trends and digital conversations. It also performs sentiment analysis on mapped data using the VADER tool [@Hutto_Gilbert_2014]. This software is a valuable resource for the scientific community, enabling the exploration of public discussions across multiple social platforms. It supports a range of research purposes, including the analysis of current discussions on global geopolitical issues. The installation of SIRITVIS is straightforward and facilitated by the pip package manager. Detailed installation instructions are available in the package's repository^[https://github.com/CodeEagle22/SIRITVIS/].
+SIRITVIS is a powerful tool designed to analyze text data from Reddit, Instagram, and other sources. It uses advanced techniques to clean data and identify main topics efficiently. The tool also includes features for visualizing data, making it easier to explore and understand the text.
+
+One unique feature of SIRITVIS is its ability to map social media posts worldwide. This means it can show where people are talking about certain topics globally, helping to understand trends and conversations happening around the world. It also analyzes the mood or sentiment of these posts using a tool called VADER [@Hutto_Gilbert_2014].
+
+SIRITVIS is very useful for researchers who want to study public discussions on social media. It is especially helpful for analyzing current events and global issues. The tool is easy to install using pip, a package manager. You can find detailed installation instructions on its GitHub page^[https://github.com/CodeEagle22/SIRITVIS/].
 
 ## Statement of Need
+Social media platforms like Reddit and Instagram have revolutionized the way we communicate, share information, and express our opinions. They are key places for public discussions, and by analyzing text data from these platforms, we can gain insights into what people are talking about, their feelings, and current trends. This information is valuable for fields like marketing, politics, and disaster response.
 
+However, dealing with the large and constantly changing amount of text data from social media is challenging. To help with this, we have developed SIRITVIS, an open-source tool for analyzing social media text. SIRITVIS uses advanced topic modeling techniques to automatically identify and extract key topics from large text datasets, making it easier to explore and find meaningful patterns.
 
-The proliferation of social media platforms has significantly changed how we communicate, share information, and express opinions on a variety of topics. Platforms like Reddit and Instagram serve as major hubs for public discussion. Analyzing text data from these platforms can provide insights into public sentiments, preferences, and trending discussions, which are valuable for areas such as marketing, politics, and disaster management.
+SIRITVIS supports several advanced topic models such as AVITM [@srivastava2017autoencoding], Latent Dirichlet Allocation (LDA) [@blei2003latent], Neural Latent Dirichlet Allocation (NeuralLDA) [@srivastava2017autoencoding], Product Latent Dirichlet Allocation (ProdLDA) [@srivastava2017autoencoding], and Contextualized Topic Models (CTM) [@bianchi-etal-2021-cross]. These models help in discovering key topics without prior knowledge of the data, making it easier to analyze and understand large collections of text.
 
[…]
