### 2024-07-22 #6243 @samhforbes
Hi @CodeEagle22 just a couple remaining,
- Terragni et al should have the proceedings all capitalised since they are in Bianchi et al (esp "European")
- Likewise the proceedings title in Sievert should also capitalise "Workshop on Interactive Language Learning, Visualization, and Interfaces" to match
