+++ paper/paper.bib
@@ -1,13 +1,8 @@
 @inproceedings{terragni-etal-2021-octis,
     title = "{OCTIS}: Comparing and Optimizing Topic Models Is Simple!",
-    author = "Terragni, Silvia  and
-      Fersini, Elisabetta  and
-      Galuzzi, Bruno Giovanni  and
-      Tropeano, Pietro  and
-      Candelieri, Antonio",
-    editor = "Gkatzia, Dimitra  and
-      Seddah, Djam{\'e}",
-    booktitle = "Proceedings of the 16th Conference of the European Chapter of the Association for Computational Linguistics: System Demonstrations",
+    author = "Terragni, Silvia and Fersini, Elisabetta and Galuzzi, Bruno Giovanni and Tropeano, Pietro and Candelieri, Antonio",
+    editor = "Gkatzia, Dimitra and Seddah, Djam{\'e}",
+    booktitle = "PROCEEDINGS OF THE 16TH CONFERENCE OF THE EUROPEAN CHAPTER OF THE ASSOCIATION FOR COMPUTATIONAL LINGUISTICS: SYSTEM DEMONSTRATIONS",
     month = apr,
     year = "2021",
     address = "Online",
@@ -17,6 +12,7 @@
     pages = "263--270",
     abstract = "In this paper, we present OCTIS, a framework for training, analyzing, and comparing Topic Models, whose optimal hyper-parameters are estimated using a Bayesian Optimization approach. The proposed solution integrates several state-of-the-art topic models and evaluation metrics. These metrics can be targeted as objective by the underlying optimization procedure to determine the best hyper-parameter configuration. OCTIS allows researchers and practitioners to have a fair comparison between topic models of interest, using several benchmark datasets and well-known evaluation metrics, to integrate novel algorithms, and to have an interactive visualization of the results for understanding the behavior of each model. The code is available at the following link: \url{https://github.com/MIND-Lab/OCTIS}.",
 }
+
 
 @inproceedings{10.1145/2484028.2484166,
 author = {Mehrotra, Rishabh and Sanner, Scott and Buntine, Wray and Xie, Lexing},
@@ -73,16 +69,13 @@ version = {1.9.1},
 year = {2023}
 }
 
+
+
 @inproceedings{sievert-shirley-2014-ldavis,
     title = "{LDA}vis: A Method for Visualizing and Interpreting Topics",
-    author = "Sievert, Carson  and
-      Shirley, Kenneth",
-    editor = "Chuang, Jason  and
-      Green, Spence  and
-      Hearst, Marti  and
-      Heer, Jeffrey  and
-      Koehn, Philipp",
-    booktitle = "Proceedings of the Workshop on Interactive Language Learning, Visualization, and Interfaces",
+    author = "Sievert, Carson and Shirley, Kenneth",
+    editor = "Chuang, Jason and Green, Spence and Hearst, Marti and Heer, Jeffrey and Koehn, Philipp",
+    booktitle = "PROCEEDINGS OF THE WORKSHOP ON INTERACTIVE LANGUAGE LEARNING, VISUALIZATION, AND INTERFACES",
     month = jun,
     year = "2014",
     address = "Baltimore, Maryland, USA",
@@ -92,6 +85,8 @@ year = {2023}
     pages = "63--70",
 }
 
+
+
 @inproceedings{bianchi-etal-2021-cross,
     title = "Cross-Lingual Contextualized Topic Models with Zero-Shot Learning",
     author = "Bianchi, Federico  and
@@ -99,7 +94,7 @@ year = {2023}
       Hovy, Dirk  and
       Nozza, Debora  and
       Fersini, Elisabetta",
[…]
