### 2024-07-23 #6243 @CodeEagle22
Hello @samhforbes,

I've updated the paper according to my understanding. Could you please review it again and let me know if everything is okay?

Thank you!
