+++ paper/paper.bib
@@ -1,8 +1,25 @@
+@inproceedings{sievert-shirley-2014-ldavis,
+    title = "{LDA}vis: A Method for Visualizing and Interpreting Topics",
+    booktitle = "{Proceedings of the Workshop on Interactive Language Learning, Visualization, and Interfaces}",
+    author = "Sievert, Carson and Shirley, Kenneth",
+    editor = "Chuang, Jason and Green, Spence and Hearst, Marti and Heer, Jeffrey and Koehn, Philipp",
+    month = jun,
+    year = "2014",
+    address = "Baltimore, Maryland, USA",
+    publisher = "Association for Computational Linguistics",
+    url = "https://aclanthology.org/W14-3110",
+    doi = "10.3115/v1/W14-3110",
+    pages = "63--70",
+}
+
+
+
+
 @inproceedings{terragni-etal-2021-octis,
     title = "{OCTIS}: Comparing and Optimizing Topic Models Is Simple!",
+    booktitle = "{Proceedings of the 16th Conference of the European Chapter of the Association for Computational Linguistics: System Demonstrations}",
     author = "Terragni, Silvia and Fersini, Elisabetta and Galuzzi, Bruno Giovanni and Tropeano, Pietro and Candelieri, Antonio",
     editor = "Gkatzia, Dimitra and Seddah, Djam{\'e}",
-    booktitle = "PROCEEDINGS OF THE 16TH CONFERENCE OF THE EUROPEAN CHAPTER OF THE ASSOCIATION FOR COMPUTATIONAL LINGUISTICS: SYSTEM DEMONSTRATIONS",
     month = apr,
     year = "2021",
     address = "Online",
@@ -14,23 +31,38 @@
 }
 
 
-@inproceedings{10.1145/2484028.2484166,
-author = {Mehrotra, Rishabh and Sanner, Scott and Buntine, Wray and Xie, Lexing},
-title = {Improving LDA Topic Models for Microblogs via Tweet Pooling and Automatic Labeling},
-year = {2013},
-isbn = {9781450320344},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-url = {https://doi.org/10.1145/2484028.2484166},
-doi = {10.1145/2484028.2484166},
-abstract = {Twitter, or the world of 140 characters poses serious challenges to the efficacy of topic models on short, messy text. While topic models such as Latent Dirichlet Allocation (LDA) have a long history of successful application to news articles and academic abstracts, they are often less coherent when applied to microblog content like Twitter. In this paper, we investigate methods to improve topics learned from Twitter content without modifying the basic machinery of LDA; we achieve this through various pooling schemes that aggregate tweets in a data preprocessing step for LDA. We empirically establish that a novel method of tweet pooling by hashtags leads to a vast improvement in a variety of measures for topic coherence across three diverse Twitter datasets in comparison to an unmodified LDA baseline and a variety of pooling schemes. An additional contribution of automatic hashtag labeling further improves on the hashtag pooling results for a subset of metrics. Overall, these two novel schemes lead to significantly improved LDA topic models on Twitter content.},
-booktitle = {Proceedings of the 36th International ACM SIGIR Conference on Research and Development in Information Retrieval},
-pages = {889–892},
-numpages = {4},
-keywords = {topic modeling, microblogs, lda},
-location = {Dublin, Ireland},
-series = {SIGIR '13}
+@inproceedings{bianchi-etal-2021-cross,
+    title = "Cross-Lingual Contextualized Topic Models with Zero-Shot Learning",
+    booktitle = "Proceedings of the 16th Conference of the European Chapter of the Association for Computational Linguistics: Main Volume",
[…]
