### 2024-08-01 #6243 @CodeEagle22
Hello @samhforbes,

The latest version of the paper is more accurate, as you requested. I urge you if you can generate the PDF again.

Thank you!
