### 2024-08-02 #6243 @samhforbes
@xillig @CodeEagle22 the PDF looks good thanks.
Can I just check the GitHub states a MIT licence, but on Figshare it's CC-BY. Can this be fixed to match please?
Likewise, we require the archive to list authors matching the paper.
