### 2024-08-04 #6243 @CodeEagle22
Hey @samhforbes, I've updated the LICENSE file in the Git repo to CC-BY with correct all authors. Looking forward to any feedback you might have!
