### 2024-08-08 #6243 @samhforbes
Hi both @CodeEagle22 and @xillig I just wanted to double check - but please note in submission requirements that JOSS needs OSI approved licences - see here: https://opensource.org/license
CC-BY is not OSI approved, but switching both to MIT or something else that is approved would work.
Otherwise agreed, I think all is looking good!
