### 2024-08-08 #6243 @CodeEagle22
Hi @samhforbes, the license has been updated to MIT on both GitHub and Figshare.
