### 2024-08-08 #6243 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
